+++ SMCL_Demeo/Properties/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle(MelonLoader.BuildInfo.Description)]
+[assembly: AssemblyDescription(MelonLoader.BuildInfo.Description)]
+[assembly: AssemblyCompany(MelonLoader.BuildInfo.Company)]
+[assembly: AssemblyProduct(MelonLoader.BuildInfo.Name)]
+[assembly: AssemblyCopyright("Created by " + MelonLoader.BuildInfo.Author)]
+[assembly: AssemblyTrademark(MelonLoader.BuildInfo.Company)]
+[assembly: Guid("E75DD5B4-4BE4-4387-80FA-79C143CC72F1")]
+[assembly: AssemblyVersion(MelonLoader.BuildInfo.Version)]
+[assembly: AssemblyFileVersion(MelonLoader.BuildInfo.Version)]
+[assembly: MelonLoader.PatchShield]